Refactor helper to separate file
